+++ module/deltagreen.js
@@ -154,30 +154,6 @@ Hooks.on("renderActorDirectory", (app, html, data) => {
   }
 });
 
-/**
- * We use this hook to translate the sample Typed Skill
- * Note - this event is fired on only the client who did the create action.
- */
-// eslint-disable-next-line no-unused-vars
-Hooks.on("preCreateActor", async (actor, creationData, options, userId) => {
-  // On brand new actors, creationData only has properties: `name`, and `type`.
-  // If creationData has `system` then the new actor is either duplicated or imported,
-  // We only want to translate the sample Typed Skill on brand new actors,
-  // thus we return early if creationData has the `system` property so we do not override anything.
-  // Also, only translate actor types with a default Typed Skill (agents and NPCs)
-  if (creationData?.system || !["agent", "npc"].includes(actor.type)) return;
-
-  // Translate the default typed skill for brand new actors.
-  actor.updateSource({
-    "system.typedSkills.tskill_01": {
-      label: game.i18n.localize("DG.TypeSkills.Subskills.Painting"),
-      group: game.i18n.localize("DG.TypeSkills.Art"),
-      proficiency: 0,
-      failure: false,
-    },
-  });
-});
-
 // Note - this event is fired on ALL connected clients...
 Hooks.on("createActor", async (actor, options, userId) => {
   try {